change: 認証ロジックをAmplifyからSupabase仕様に載せ替え
